Make the OutputGUI's log method always put the cursor at the end of the document
We previously made it so that the OutputGUI's log method always sets the cursor at the beginning of the document.  This isn't that nice since the side-effect is that the document gets scrolled to the end.  Let's keep the cursor at the end so that users will see the end of the log (which is where the most recent/useful information is found).

Signed-off by: David Aguilar <davvid@gmail.com>