Make the historybrowser config setting show on restart
Renamed the historybrowser config name to match the other config names.
Added the historybrowser parameter to the allowed list so that it is serialized into git config.
Added the historybrowser param to the refreshview method of the optionscontroller so that it is properly read on startup.

Signed-off by: David Aguilar <davvid@gmail.com>