Change the qobserver's notification order.  This fixes a bug and removes code made unnecessary by this design.
We now change the disabling of model notification in when setting Qt widget params instead of when setting model params.  This undoes a previous fix that was added to the options controller that would refresh the historybrowser line in the options gui.  That code is no longer needed with this change.  An added benefit is that the interactive font updating is fixed/enabled with this fix.

Signed-off by: David Aguilar <davvid@gmail.com>