Add a gui for displaying arbitrary commits/blobs/trees
Added Repository->Show Revision

This menu item lets you search for revisions using a text input.  The first revision matching the query string is displayed in the gui.  If the sha1 doesn't match any commits, we look for a blob or tree that matches the sha1 and display it.

Signed-off by: David Aguilar <davvid@gmail.com>