Set window sizes and default fonts
If no font is specified for the font combo box then no font will be listed for
the diff font.  Set a selected font so that the field is always populated.
This only affects the case where a user has never previously selected a font.

Signed-off by: David Aguilar <davvid@gmail.com>