Properly undo changes when cancelling out of the options gui using the ESC key
Previously, hitting ESC to exit out of the options gui was not properly notifying
the main app about needing to update/revert the fonts back to their original values.
This fixes that problem and removes some code by specifying the connection between
the optionsgui.cancel_button signal and the optionsgui.reject slot in the ui file.
We then catch the reject signal in the controller and handle that.  The reason it
didn't work previously is that we were only handling the case were cancel was pressed
instead of explicitly handling the rejected() signal.

Signed-off by: David Aguilar <davvid@gmail.com>