git.py: use a dispatcher to allow for more dynamism
The goal of these changes is to allow runtime creation
of arbitrary git commands.  e.g. git.branch() should
return the output of "git branch."  We do this by
using a module dispatcher/__getattr__ to intersept
attributes.  Eventually, the git.py module should
be very tiny as the hand-coded methods are
phased out.

Signed-off-by: David Aguilar <davvid@gmail.com>