cola: add mergetool support
git-cola now has a mergetool option in the
right-click menu when running on non-macos,
non-windows platforms.

The reason for blocking out windows and macos
is that there is no simple way to spawn a terminal
for the interactive mergetool session.

We could, of course, write our own mergetool wrapper
in Python, but this is the easiest way for now.

Signed-off-by: David Aguilar <davvid@gmail.com>