ui: redo the layout so that we have a central widget
Previously, the main interface shunned having a central
widget in favor of purely having dock widgets in the
dock areas.  Since I haven't figured out how to make
a splittable dock area be the central widget, I made
the status widget the central widget and placed all of
the other windows in the left dock area.

TODO: figure out whether we can have a nice splittable,
dockable layout.  It was possible in qt3, but it does seem
like some of this functionality might not be accessible
from python.

Signed-off-by: David Aguilar <davvid@gmail.com>