git-difftool: compare against the index by default for all commands
This reworks git-difftool so that the command invocations are slightly
closer to git-mergetool.  One big difference is that we show things from
oldest to newest when there are changes in both the worktree and the index.

In that case:
	BASE=the latest commit
	LOCAL=the index
	REMOTE=the worktree

Certain diff commands were adjusted so that they would show these three
states in a left-to-right progression since that made most sense to
english speakers.

Signed-off-by: David Aguilar <davvid@gmail.com>