cola: completely rework the cola user interface
git-cola now provides a repository status view that closely matches the
output of 'git status'.

The staged and unstaged lists were replaced by a single tree view.
Likewise, rest of the UI was simplified to account for this --
e.g. there are no more dock widgets in use.

New icons were added for the Staged, Modified, and Unmerged categories.

Signed-off-by: David Aguilar <davvid@gmail.com>