Guess the correct encoding for each diff line.
Git can handle any type of file, whatever encoding it uses
(or a broken one). Git-cola must not fail (or complain) when
trying to show that data.

Git-cola only handles utf8 in diffs, which is not very useful
when your project has different encodings.

This patch adds a very simple encoding guess for each line.
Do not try to do this at the whole file, because it is useful
when you are viewing a diff of a encode change.

Signed-off-by: David Martínez Martí <deavidsedice@gmail.com>