search: fix bug that blocks searching
When trying to search something, git-cola complains:
 > AttributeError: Parameter 'commit_list' not available for Model

Attributes for class "Model" must not be null to be detected as a parameter.
Model.get_param_names has the related code to check if it is a parameter
or not.

So, the solution here is initialize model.commit_list with an empty array,
because it is similar to the data used there.

Signed-off-by: David Martínez Martí <deavidsedice@gmail.com>