interactive staging: workaround a qt quirk where \r\n is lost in qtextedits
qtextedits unconditionally eat \r\n when returning the selection.
This breaks the interactive staging on files that contain \r\n.
If we're not able to find the selection in the diff then we know
that this has happened so simply replace \n with \r\n and try
again.

This fixes interactive staging on repositories that contain
windows/dos text files.

Signed-off-by: David Aguilar <davvid@gmail.com>