clone: assume that git url = the new destination directory
Previously we used QFileDialog.getSaveFileName() and restricted
users from selecting existing directories.

We now use getExistingDirectory and assume that the destionation
directory should just be the basename of the git url.

We could prompt again, but that would be annoying.  Unfortunately
Qt doesn't have a convenient dialog to allow both existing
directories and new directory names (most of the world is centered
around files, it seems, unlike git which is more dir tree-ish).

Signed-off-by: David Aguilar <davvid@gmail.com>