qobserver: fix a bug where we would respond to redundant notifications
When setting a text field Qt triggers a callback.  We catch this callback
and update our model accordingly.  When the model updates, it triggers
another callback which causes a feedback loop that eventually resolves
with the text field getting set twice.  This is problematic since
it makes entering text problematic.

We now detect this situation and stop the notification loop accordingly.

Signed-off-by: David Aguilar <davvid@gmail.com>