win32: vastly simplify the installation procedure for Windows
We now automatically install the bin/cola win32 wrapper script.
This is now the only recommended way to run cola.  Installing
cola on Windows is now no more difficult then unzipping a tarball
and running the 'cola' shell script.

The 'cola' wrapper script is now smarter in that it checks for
both Python2.5 and Python2.6, thus cola can now run with either
version.

Signed-off-by: David Aguilar <davvid@gmail.com>