git: change with_extended_output to only return status, output
The stderr+stdout handling in the git module was broken because it
could deadlock on older versions of Python.  All of the complication
came from trying to handle both stderr and stdout in the same
function.  We now pass stderr=subprocess.STDOUT to subprocess.Popen
so that things become much simpler.  We only include stderr when
with_extended_output=True, otherwise we ignore it.

This fixes deadlocks on older Pythons and on Mac OSX.
It also seems to speed things up a bit on Linux, which is a plus.

Signed-off-by: David Aguilar <davvid@gmail.com>