Detect upstream path changes on tracking branches
In the classic view, when the current branch is tracking a remote
branch, detect when there are changes to paths on the tracked branch.
The 'Changed Upstream' status is lower priority than local changes
(e.g., if a file is changed upstream, but is also locally modified,
the classic view will show it as locally modified, but not changed
upstream, since that is probably just confusing and there is limited
space in the status area).

A corresponding new action to 'checkout updated file' should be added
to the list of actions for paths, and enabled when the path is
'changed upstream'.

Signed-off-by: Uri Okrent <uokrent@gmail.com>