qtutils: Make SLOT take optional arguments
If callers do not pass in any values when SLOT is called then we
relay the values passed in by the callback invoker.  When values
are provided then we prefer those values over the ones available
when the callback is invoked.

This makes it so that values provided at registration-time
override run-time values from the signal.

Signed-off-by: David Aguilar <davvid@gmail.com>