models.gitrepo: Ignore unknown paths
PyQt4 segfaults when children are added to QStandardItemModels
after a view is observing it.

This works on older version of PyQt4 so this is a workaround
until we have a better solution.

The downside of this change is that we won't recognize when
new files appear in the repository.  The upside is that it
avoids PyQt4 segfaults.

Closes #25

Signed-off-by: David Aguilar <david.aguilar@disneyanimation.com>