Make our diffs more patient -- patience is a virtue
Make --patience the default for diffs produced by cola, if the git
version supports it (that is the user's git client is at least 1.6.2).

Maybe this should be controlled by a configuration somewhere, but I'm
too lazy to make that work now. Also, git config doesn't seem to mention
any existing config for controlling whether or not diff is patience by
default so it doesn't seem like cola should be the thing deciding that.
Is there any downside to always using patience diff?

Signed-off-by: Uri Okrent <uokrent@gmail.com>