Use the new gettext infrastructure for i18n
This adds a cola.i18n module for doing gettext translations.
This does away with the qm-based translation system in favor
of a standard gettext setup.

We still monkey-patch QApplication.translate() with our
own custom translator.  The main difference is switching
from a Qt-specific QTranslator translation system to a
gettext-based one.

Signed-off-by: David Aguilar <davvid@gmail.com>