cmdfactory: Add 'prompt_user' for delegating user interaction
The command factory now allows us to prompt for user inputs.
This is done through a "Command Wrapper" class that is installed
for each parameter that needs user interaction from a command.

This allows the commands to be decoupled from their GUI
components.  The dependencies are injected at runtime by
the GUI.  Once the wrappers are in place then commands
can prompt for input by going through the command factory.

Signed-off-by: David Aguilar <davvid@gmail.com>