difftool: running on modified+staged path shouldn't show staged changes
Running difftool on a path that is both modified and staged can be done
from two spots in the repo status tool, the staged version of the path
and the modified version of that path.  The difftool run on the staged
version was correctly only showing the staged changes, but run on the
modified section was showing both the staged and modified changes, when
it should really leave out the staged changes.

Signed-off-by: Uri Okrent <uokrent@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>