views.dag: Improve the graph layout algorithm
This algorithm is slightly more appealing because it bends chains
of commits that lead to parallel branches.

It's not perfect, but it's one of the simpler things we can do
with the streaming one-node-at-a-time approach in use.

We could probably make a very nice layout by post-processing
the graph using a thread but it's not *that* bad, either.

Signed-off-by: David Aguilar <davvid@gmail.com>