mainwindow: Simplify the 'Actions' widget
PyQt4 4.7.4 loses its bearings when a widget modifies the
direction of its QBoxLayout in a resize event.

We depended on this behavior to allow the actions widget to
automatically switch between vertical and horizontal layouts.

Simplify the code by just sticking to a vertical layout.
All of the actions on the 'Actions' widget are also
available in the menus these days.

Closes #62

Reported-by: alyst via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>