darwin: Remove dependency on py2app for OS X deployment
We now ship separate installers for Qt/PyQt so we don't need to
ship them inside of git-cola.app anymore.  This cuts down on the
OS X download size tremendously and allows us to ship a single
git-cola.app for all OS X platforms.

The only difference between platforms are the Qt/PyQt libraries
which are available as a separate download.

Signed-off-by: David Aguilar <davvid@gmail.com>