inotify: Rework thread timeouts to consume less resources
Register known paths up front to avoid checks within a loop.
Updates timeouts so that we sleep for slightly longer.
We no longer use quit() when ending the inotify thread.
We call stop() and wait for it to complete.  This allows us
to do without the setTerminationEnabled(True) call,
which is generally discouraged.

Signed-off-by: David Aguilar <davvid@gmail.com>