views: Make observer callbacks thread-safe
Prepare to offload expensive operations onto a background thread
by making all of the observer-registered GUI callbacks thread-safe.
We do this by using Qt's signal/slots.  When a callback happens in
another thread it simply posts a signal using emit().

The main (GUI) thread registers for these signals and handles
updating the views in response to the signals.

Signed-off-by: David Aguilar <davvid@gmail.com>