utils: Fix launching background commands on windows
Newer versions of msysgit no longer require the win32_expandpaths()
hack to expand paths after the double-dash.

Remove this old hack to make newer gits happy.

This basically means we don't support the older msysgit which
required the workaround, which is fine.  If we find someone is
still using an old version we can always add a check later.

Signed-off-by: David Aguilar <davvid@gmail.com>