main: Add a '--prompt' option and use it in the desktop launchers
git-cola assumes that the current directory is a git repository.
This is less helpful when launching git-cola through a launcher icon
because it may find a git repository in $HOME.

Add a '--prompt' flag and use it in all of the launchers.
This forces git-cola to prompt the user for a repository before
launching the main GUI.

Closes #75

Signed-off-by: David Aguilar <davvid@gmail.com>