gitcmds: Fix unstaging of files when amending a commit
Pass an explicit head when unstaging files with 'git reset'.
This makes unstaging files from the previous commit work correctly
since we do not always want to reset to the contents of 'HEAD'.

The net effect is that we now call 'git reset HEAD -- <paths>' for the
normal case and 'git reset HEAD~ -- <paths>' when amending a commit.

Closes #82

Signed-off-by: David Aguilar <davvid@gmail.com>