perf: Avoid calling 'git update-index --refresh' too often
We now call 'git update-index --refresh' on startup and in the 'Refresh'
command only.

The other commands call model.update_file_status() which avoids
updating the entire index, updating remotes, etc.

This shaves off a few unneeded git calls.  The stage/unstage methods
previously included some performance hacks to avoid needing to rescan
for changes.  These have been removed in favor of the simpler "just rescan"
logic.  This will become less of an issue in the future when we do all
rescanning in a background thread.

An important note is that all of the editor commands call update_status()
explicitly instead of relying on the base class to call it for them.
This is better since it makes things more explicit.

Signed-off-by: David Aguilar <davvid@gmail.com>