Makefile: Better support for out-of-git builds
The top-level Makefile depended on the existence of cola's
git repository to generate its version number.

Rework the Makefile so that the "make dist"-generated "version"
file is used instead.  This is done by scraping the output of
"python cola/version.py", which knows to look for the "version"
file before trying to use "git describe".

Signed-off-by: David Aguilar <davvid@gmail.com>