remote: Improve user interaction for remote actions
Make sure the 'remote' section is always at the top and focused.
When doing 'pull', do not select a local branch by default.

This ensures that we do "git pull $remote" instead of
"git pull $remote $branch", which only updates the local
branch and leaves the remote tracking branch behind.

Signed-off-by: David Aguilar <davvid@gmail.com>