dag: Make labels into a separate item class
The label drawing code was originally part of the Node item class.
This had the negative side-effect that clicking on the label would
select the node even when the node contained no labels.

Use a separate, non-selectable, QGraphicsItem subclass for labels.
Make it a child of the node so that it can be positioned relative
to the node's position in the scene graph.

Signed-off-by: David Aguilar <davvid@gmail.com>