packaging: Depend on git >= v1.6.3
git-difftool originated in git-cola and was later contributed to
the main git project's contrib/ section.  It moved out of contrib/
in git v1.6.3.

git-cola also uses "git mergetool --no-prompt" and "git diff --patience"
which both appeared in v1.6.2.

Remove our version of difftool and depend on git >= v1.6.3 so that
we do not have to maintain these scripts anymore.  Remove the runtime
checks for "mergetool --no-prompt" and "diff --patience" now that the
entire project depends on a modern git.

Signed-off-by: David Aguilar <davvid@gmail.com>