views.main*: Simplify menus, reduce code, and fix Ctrl+C in diff viewer
Remove the unnecessary 'Edit' menu from the main menu bar.
The 'Options' action moved into the 'File' menu since the
'Edit' menu is empty without Cut/Paste/etc.

Centralize all of the action handling code into views.mainwindow.
Do not register any actions for Copy/Paste/etc. whatsoever so that
the Qt defaults (which do the right thing) are used instead.

This is the simpler thing to do, uses less code, makes a nicer UI,
and fixes being able to use Ctrl+C in the diff viewer.

It also reduces confusion because it is not clear for new users whether
the edit menu operates on the commit message editor or on the diff viewer.

Likewise, leaving the Ctrl+C, etc. keyboard shortcuts alone allows
Qt to use the correct widget-context action (which happens by default).

This is a net reduction of ~100 lines of code. yay :-)

Closes #90

Signed-off-by: David Aguilar <davvid@gmail.com>