Makefile: Remove extra work done in "make dist"
Our tarballs do not need to include the extra "version"
file anymore so remove the unnecessary steps from the dist target.

This target won't really be used anymore since we are just
publishing the repo as-is without any post-processing of the
output of "git archive", but it's still nice to have.

Signed-off-by: David Aguilar <davvid@gmail.com>