gitcmds: Simplify, speedup, and refactor worktree_state()
We originally ran "git diff --exit-code <path>" against every
modified file to detect "partially-staged" files.

Simplify the code so that we only ever call two git commands to
gather staged and modified files.  This eliminates one git call
on startup in the best case and many calls in the worst case
when many files have been modified.

We may later want to split out the diff_upstream(head) call so
that it is invoked by the classic view only since it is not used
anywhere else.

Signed-off-by: David Aguilar <davvid@gmail.com>