widgets.commitmsg: Shift+tab to focus summary
It would be nice if `tab` went to the description editor like `return`
but that would make it impossible to shift to the 'Commit' button.

`Shift+tab` from the description field to the summary field is
convenient even though it is not symmetric to the `tab` behavior.
The convenience outweighs the inconsistency so make doing a
`Shift+tab` from the description field focus the summary field.

Signed-off-by: David Aguilar <davvid@gmail.com>