widgets.diff: Do not resize text on wheel scroll + control
Create a new wheel event without the modifier to get a nice
scrolling behavior instead of a problematic resize.

The resize that happens in the normal wheel event is bad because
the font resize does not persist and it is not obvious how to revert
back to the default diff font size.

Signed-off-by: David Aguilar <davvid@gmail.com>