main: Simplify the "Diff->Expression" action
Use a modal dialog instead of changing the mode of the main
interface when diffing using an expression.

This uses the difftool module to diff using an expression and
eliminates the special "diff mode".

The "diff mode" was a usability problem because the "Exit Diff Mode"
button only ever appeared in the "Actions" widget.

Using a dedicated dialog makes things simpler and easier to use.

Signed-off-by: David Aguilar <davvid@gmail.com>