dag: Keep selection when showing context menus
Doing control+click in the list view shows a context menu but it
also resets the selection.  This makes the "Diff selected..."
actions useless since they rely on the existing selection.

Ignore the mouse press event when the modifier is Meta,
AKA the "control" button on OS X.  Linux users already had
this tweak when applied for the right mouse button.

Signed-off-by: David Aguilar <davvid@gmail.com>