widgets.commitmsg: Jump to the next line only when at EOL
Hitting 'Down' in the summary field is still useful for jumping
to the end of the summary.

Make the 'Down' key work by first jumping to the end of the line
(which is the default behavior we get for free from Qt) and then
a second down (when at the end of the line) jumps to the
description field.  This makes the summary field act just like
the description field which is another usability plus.

Signed-off-by: David Aguilar <davvid@gmail.com>