classic: Make the title match 'cola' and 'dag'
Make the titles consistent for now.  These are different tools,
though, so we might want to be able to differentiate them one day.

We might want to make the titles say this eventually:

	cola: <project>: <branch>
	dag: <project>: <branch>
	browser: <project>: <branch>

..but for now we omit the "tool: " prefix to save panel real estate.

Having different logos might be a better way in the future instead
of changing the titles as well.

Signed-off-by: David Aguilar <davvid@gmail.com>