browser: Rebrand the 'Classic' tool as 'Browser'
The 'Classic' tool was slightly inspired by classical "file-based"
browsers such as tkcvs and provides an entire view of the repo which
is not available in the focused 'Status' widget.  Likewise, it was
kind of funny to say `git cola classic`.

Inside jokes aside, it's better to rebrand this tool as the 'Browser'
since browsing is what it does.  The next thing it really needs is
a vim-like "/" search function to filter (or possibly just find) paths.
Being able to "chdir" into a subdirectory and see its contents only
would be a nice as would dedicated list, item, and tree modes.

Signed-off-by: David Aguilar <davvid@gmail.com>