dag: Add gravatar icons and an elegant new design
Remove the author and summary fields from the display and
use dedicated QLabels to display them.  The diff displays
the extended description only.

Later work would be adding links to parents and children.
A corner widget menu might be appropriate.

Gravatar image lookups are done asynchronously and
use an internal pixmap when an image does not exist.

Signed-off-by: David Aguilar <davvid@gmail.com>