dag: Properly handle right-clicking in an empty area
update_menu_actions() assumed that there was always a clicked item
and thus we were trying to use None as a commit.

Add safety checks to avoid this problem and make it so that the 'copy'
action can be performed even when the right-click action happens in
an empty area.

Signed-off-by: David Aguilar <davvid@gmail.com>