dag: make branches easier to follow by coloring edges
Change the edge color whenever an edge marks the start of a new branch
and continue using that edge color for children on it.  This makes
branches easier to follow.

The algorithm used to detect a new branch is pretty brain-dead.  An edge
is on a new branch if the edge is longer than the standard edge between
a parent and child.

Signed-off-by: Uri Okrent <uokrent@gmail.com>