Only change edge color on new branch creation
This makes our edge colors a bit less schitzophrenic by only switching
the edge color when a new branch is detected, rather than also when the
branch is merged in.  This makes the history a bit easier to follow.

It also prevents the colors from changing when multple merges are
integrated from the same lineage (such as when changes are being
continually merged in from a stable branch, the stable branch will
maintain it's edge color through repeated merge-backs, and all the
merge-back edges will be the same color as the stable branch).

Signed-off-by: Uri Okrent <uokrent@gmail.com>