win32: Avoid super() everywhere for portability
Python 2.6 on win32 has a super() that does not always work.
It could be the PyQt4 version.  I don't know.

It sometimes works, it sometimes doesn't.  It seems like it
doesn't work when called from within a class created by a
closure (which we do a lot).

Rather than try and pick and choose, just avoid super()
altogether since for portability and consistency.

Signed-off-by: David Aguilar <davvid@gmail.com>