cmds: Teach the Edit command lots of new tricks
notepad++, textedit, and emacs users gained the ability to jump
to a specific line when editing a file listed in a grep result.

The edit command now works on multiple files, making it possible
to launch an editor with many files at a time.

The gui.editor variable is now parsed with the shell, which
is useful if you have something like 'gvim -p' or 'gvim -o'
as the gui.editor now that we open many files at once.

Signed-off-by: David Aguilar <davvid@gmail.com>