commitmsg: Add 'Sign Off' and 'Commit' to the context menus
We recently made it so that 'tab' jumps from the summary to the
extended description.

A consequence of this is that we can no longer tab over to the
commit button.  Add a new entry point to this action in the
summary and extended description's context menus to compensate
for it.

This also advertises the 'Ctrl+Return' keyboard shortcut in the
context menu, which makes it more discoverable.

Signed-off-by: David Aguilar <davvid@gmail.com>