dag: Some graph drawing enhancements
We now detect the true height of the tag labels.
We might later want to use different labels for local
branches (green), remote branches (tan), and tags (yellow).

The font was made slightly larger and the node sizes and
margin were adjusted to provide a large enough margin when
zooming with a mouse wheel.

Signed-off-by: David Aguilar <davvid@gmail.com>