diff: Support .gitattribute encodings when showing diffs
Teach the diff viewer to display diffs using the encoding
specified in .gitattributes.

This is the first half of the work needed to properly support
.gitattributes for file encodings as described in issue #96.

Set `cola.fileattributes` to true to enable this feature.
The diffs are now displayed correctly for non-utf8 files.

Diffs cannot yet be applied, but this is the first step.

Signed-off-by: David Aguilar <davvid@gmail.com>