usability: Flip some prompt defaults for more intuitive interaction
Prompts that are initiated by the user, e.g. "Revert Section", should
default to True.  Otherwise, the user has to needlessly click when they
could more easily press "Enter".

This is the way it should have been from the beginning.  The initial
menu click is enough of an indicator that the user wants to do
something.  The follow-up question should therefore default to True.

Signed-off-by: David Aguilar <davvid@gmail.com>