utils: Do not word-wrap sign-offs and special tags
Teach the word-wrapper to skip over lines that look like sign-offs,
which is basically any line that starts with a string of letters and
dashes followed by a colon.  Other popular tags are things like
"Acked-by:", "Helped-by:", etc.  We support all of these.

If we're going to have our own word-wrapping function then we
might as well do something useful with it ;-)

Signed-off-by: David Aguilar <davvid@gmail.com>