status: Use nicer icons and make the header items more distinct
Save screen real-estate by eliminating the needless indentation which
was being applied to the left side of items in the status view.

Eliminate the icons used by the status header items since they are no
longer needed.  Consolidate the icons so that the "sigil" ones are used
in the status view.

Add a 2-pixel spacing between items to avoid cluttering things too
tightly.

Signed-off-by: David Aguilar <davvid@gmail.com>