darwin: Trick OS X into saying "git-cola" in the menu bar
Our .app bundle has always said "Python" in the menubar and "git-cola"
in the task switcher.  We use CFBundleName in the Info.plist to set the
application name, but the menubar name is always the name of the current
process, which for PyQt4 apps is typically "Python".

OS X can coerced into showing a different name in the title by launching
python through a named symlink which points to the system framework's
Python.app.  It does not work with /usr/bin/python because it escapes
the bundle.

Create symlinks during "make git-cola.app" so that a Resources/git-cola
points to the system's framework Python.app.  Adjust the launcher script
to prefer launching through the symlink.

This makes the .app bundle behave identically to a native OS X app.

Signed-off-by: David Aguilar <davvid@gmail.com>