status: Show untracked files in the background
Large binary files can exist in the worktree.
Keep the GUI from blocking by doing it in a thread.

Limit the size of the read calls to 2k to minimize
blocking the GUI.  Even though we use a thread,
python's GIL starts blocking, so this helps keep
things responsive despite the fact that we truncate
the output.

The `cola.readsize` configuration variable can be
used to adjust the limit in case 2k is not enough.

Closes #142

Signed-off-by: David Aguilar <davvid@gmail.com>