settings: Provide default values to simplify the implementation
We were catching KeyError in a number of places to account for the fact
that the user's ~/.config/git-cola/settings json file may have missing
keys.

Instead of doing that, provide a default dictionary that we update with
the contents of that file.  This allows us to safely assume that the
keys are always available.

Signed-off-by: David Aguilar <davvid@gmail.com>