version: be more robust when parsing Git versions
We cannot assume that the last element is the git version
string.  Apple's build of git, for example, adds an extra
"(Apple build XX)" to the end of the version string.

Make the parsing more robust by splitting like we did
before, but taking the 3rd element instead of assuming
that the last element is the one that contains the
version string.