gitcmds: add diff() function
Return the filenames that differ given the provided arguments.
This is an interface on top of "git diff", which is plumbing,
but we call it with "--name-only -z" which makes the output
predicatable.  This lets us have the ability to diff against index,
worktree, and arbitrary commits using a single interface.

Signed-off-by: David Aguilar <davvid@gmail.com>