cmds: ignore inotify updates when staging/unstaging
Staging operations with thousands of items can take a while, and inotify
may send updates while they are running.  The GUI already updates at the
end of the operation so there's no need to process the inotify events
while staging.

Add a CommandDisabled context manager to disable any command inside of a
"with" block.  Use CommandDisabled to ignore the UpdateFileStatus
command while we are inside a lengthy stage/unstage command.

Reported-by: Christian <horschi@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>