status: preserve horizontal scroll across updates
When a file name is too long, a horizontal scroll bar appears in the
status widget. As the selection is updated, QTreeView tries to put the
first column back in 'full view' and thus resets the scroll bar
horizontal position to "0" (source
http://doc.qt.digia.com/4.6/qabstractitemview.html#autoScroll-prop)

This commit disables that behaviour.
'setAutoScroll(..)' is available as early as in Qt 4.0.