push/pull: do not show the "OK" dialog when successful
The "git push succeeded" confirmation dialog was added because
one user was having "Push" hang on them and I wasn't sure if
they were confused because they never got a confirmation.

Go back to the original way the code worked where it only
shows you a message in the event of an error.

Closes #184

Signed-off-by: David Aguilar <davvid@gmail.com>